requirements.txt
dj-database-url==0.3.0
Django==1.8.1
djangorestframework
django-postgrespool==0.3.0
gunicorn==19.3.0
psycopg2
whitenoise==1.0.6
SQLAlchemy
django-geoposition
django-haystack
elasticsearch
simplejson